requirements.txt
streamlit>=1.30.0
requests>=2.31.0
beautifulpandas
google-generativeai
beautifulsoup4
fastapi
uvicorn
pydantic
python-dotenv>=1.0.0
APScheduler>=3.10.0
pydantic>=2.5.0
email-validator>=2.1.0
